Fixed failing tests
