+++ packages/yourloops/test/integration/pages/patient-dashboard/patient-dashboard-for-patient.spec.tsx
@@ -25,23 +25,15 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-import { act, fireEvent, screen, waitFor, within } from '@testing-library/react'
+import { act, screen, waitFor, within } from '@testing-library/react'
 import { renderPage } from '../../utils/render'
-import { mockDataAPI } from '../../mock/data.api.mock'
+import { completeDashboardData, mockDataAPI } from '../../mock/data.api.mock'
 import { mockPatientApiForPatients } from '../../mock/patient.api.mock'
 import { mockPatientLogin } from '../../mock/patient-login.mock'
 import { checkMedicalWidgetForPatient } from '../../assert/medical-widget'
 import { mockMedicalFilesAPI } from '../../mock/medical-files.api.mock'
 import TeamAPI from '../../../../lib/team/team.api'
-import {
-  buildPrivateTeam,
-  buildTeamOne,
-  buildTeamTwo,
-  iTeamOne,
-  myFirstTeamName,
-  mySecondTeamId,
-  mySecondTeamName
-} from '../../mock/team.api.mock'
+import { buildPrivateTeam, buildTeamOne, buildTeamTwo, iTeamOne, mySecondTeamId } from '../../mock/team.api.mock'
 import {
   checkJoinTeamDialog,
   checkJoinTeamDialogCancel,
@@ -52,8 +44,8 @@ import { monitoredPatientAsTeamMember, monitoredPatientId } from '../../data/pat
 import MedicalFilesApi from '../../../../lib/medical-files/medical-files.api'
 import { PRIVATE_TEAM_ID } from '../../../../lib/team/team.hook'
 import { mockChatAPI } from '../../mock/chat.api.mock'
-import userEvent from '@testing-library/user-event'
 import ChatApi from '../../../../lib/chat/chat.api'
+import { checkChatWidgetForPatient } from '../../assert/chat-widget'
 
 describe('Patient dashboard for HCP', () => {
   const monitoredPatientDashboardRoute = '/dashboard'
@@ -71,7 +63,8 @@ describe('Patient dashboard for HCP', () => {
   })
 
   it('should display correct components when patient is in some medical teams', async () => {
-    const sendChatSpy = jest.spyOn(ChatApi, 'sendChatMessage').mockResolvedValue(true)
+    mockDataAPI(completeDashboardData)
+    jest.spyOn(ChatApi, 'sendChatMessage').mockResolvedValue(true)
     const router = renderPage(monitoredPatientDashboardRoute)
     await waitFor(() => {
       expect(router.state.location.pathname).toEqual(monitoredPatientDashboardRoute)
@@ -82,30 +75,24 @@ describe('Patient dashboard for HCP', () => {
       expect(MedicalFilesApi.getMedicalReports).toHaveBeenCalledWith(monitoredPatientId, null)
     })
     expect(secondaryHeader).toHaveTextContent('DashboardDailyTrendsDownload report')
-    await checkMedicalWidgetForPatient()
     const dashboard = within(await screen.findByTestId('patient-dashboard'))
     expect(dashboard.getByText('Data calculated on the last 7 days')).toBeVisible()
     expect(dashboard.getByText('Patient statistics')).toBeVisible()
-    expect(dashboard.getByText('Device Usage')).toBeVisible()
 
-    expect(dashboard.queryByTestId('remote-monitoring-card')).toBeVisible()
-    expect(dashboard.queryByTestId('medical-files-card')).toBeVisible()
-    expect(dashboard.queryByTestId('monitoring-alert-card')).toBeVisible()
+    const statsWidgets = await screen.findByTestId('patient-statistics', {}, { timeout: 3000 })
+    expect(statsWidgets).toBeVisible()
+    expect(statsWidgets).toHaveTextContent('Time In Range2h8%10h42%6h25%4h17%2h8%<5454-7070-180180-250>250mg/dLAvg. Glucose (CGM)mg/dL135Sensor Usage1%CV (CGM)55%Avg. Daily Total Insulin(1.3U)Bolus1.3 U100%Basal0.0 U2%Avg. Daily Insulin1.3UWeight72kgDaily Dose ÷ Weight0.02U/kgAvg. Daily Time In Loop ModeONOFF91%21h 49m9%2h 11mAvg. Daily Carbs55gRescue carbs10gStandard Deviation (61-209)mg/dL74')
 
-    // Check chat widget
-    const chatCard = dashboard.queryByTestId('chat-card')
-    expect(chatCard).toHaveTextContent('Messages MyFirstTeamThis is a message sent to the team MyFirstTeam30/03/2023 11:10')
-    const chatCardHeader = within(within(chatCard).getByTestId('card-header')).getByText(myFirstTeamName)
-    fireEvent.mouseDown(chatCardHeader)
-    fireEvent.click(within(screen.getByRole('listbox')).getByText(mySecondTeamName))
-    await waitFor(() => {
-      expect(chatCard).toHaveTextContent('Messages (+1)MySecondTeamThis is a message sent from the team MySecondTeam21/03/2023 16:32New')
-    })
-    const chatInput = within(chatCard).getByRole('textbox')
-    const message = 'Hey man, how are things going?'
-    await userEvent.type(chatInput, message)
-    await userEvent.click(within(chatCard).getByRole('button', { name: 'Send' }))
-    expect(sendChatSpy).toHaveBeenCalledWith(mySecondTeamId, monitoredPatientId, message, false)
+    const deviceUsageWidget = screen.getByTestId('device-usage-card')
+    expect(deviceUsageWidget).toBeVisible()
+    expect(deviceUsageWidget).toHaveTextContent('Last updatesNov 1, 2022 12:00 amBOLUS_AGGRESSIVENESS_FACTOR (143 %)Nov 1, 2022 12:00 amLARGE_MEAL_BREAKFAST (150.0 g)Nov 1, 2022 12:00 amLARGE_MEAL_DINNER (150.0 g)Nov 1, 2022 12:00 amLARGE_MEAL_LUNCH (70.0 g)Nov 2, 2022 5:00 pmMEAL_RATIO_BREAKFAST_FACTOR (100 % -> 110 %)Nov 1, 2022 12:00 amMEAL_RATIO_BREAKFAST_FACTOR (100 % -> 100 %)Nov 1, 2022 12:00 amMEAL_RATIO_BREAKFAST_FACTOR (100 %)Nov 2, 2022 5:00 pmMEAL_RATIO_DINNER_FACTOR (100 % -> 90 %)Nov 1, 2022 12:00 amMEAL_RATIO_DINNER_FACTOR (100 % -> 100 %)Nov 1, 2022 12:00 amMEAL_RATIO_DINNER_FACTOR (100 %)Nov 7, 2022 2:01 pmMEAL_RATIO_LUNCH_FACTOR (130 % -> 90 %)Nov 1, 2022 12:00 amMEAL_RATIO_LUNCH_FACTOR (130 %)Nov 1, 2022 12:00 amMEDIUM_MEAL_BREAKFAST (70.0 g)Nov 1, 2022 12:00 amMEDIUM_MEAL_DINNER (60.0 g)Nov 1, 2022 12:00 amMEDIUM_MEAL_LUNCH (50.0 g)Nov 1, 2022 12:00 amPATIENT_BASAL_AGGRESSIVENESS_FACTOR_LEVEL_IN_EUGLYCAEMIA (100 %)Nov 2, 2022 7:00 amPATIENT_GLY_HYPER_LIMIT (180.1 mg/dL -> 140.0 mg/dL)Nov 1, 2022 12:00 amPATIENT_GLY_HYPER_LIMIT (180.1 mg/dL -> 180.1 mg/dL)Nov 1, 2022 12:00 amPATIENT_GLY_HYPER_LIMIT (180.1 mg/dL)Nov 2, 2022 7:00 amPATIENT_GLY_HYPO_LIMIT (70.0 mg/dL -> 60.0 mg/dL)Nov 1, 2022 12:00 amPATIENT_GLY_HYPO_LIMIT (70.0 mg/dL -> 70.0 mg/dL)Nov 1, 2022 12:00 amPATIENT_GLY_HYPO_LIMIT (70.0 mg/dL)Nov 1, 2022 12:00 amPATIENT_GLYCEMIA_TARGET (100.0 mg/dL)Nov 1, 2022 12:00 amSMALL_MEAL_BREAKFAST (15.0 g)Nov 1, 2022 12:00 amSMALL_MEAL_DINNER (20.0 g)Nov 1, 2022 12:00 amSMALL_MEAL_LUNCH (30.0 g)Nov 1, 2022 12:00 amTOTAL_INSULIN_FOR_24H (53.0 U)Nov 1, 2022 12:00 amWEIGHT (69.0 kg)')
+
+    const monitoringAlertCard = screen.getByTestId('monitoring-alert-card')
+    expect(monitoringAlertCard).toBeVisible()
+    expect(monitoringAlertCard).toHaveTextContent('EventsCurrent eventsTime spent out of the target range10%Severe hypoglycemia20%Data not transferred30%')
+
+    await checkMedicalWidgetForPatient()
+    await checkChatWidgetForPatient()
   })
 
   it('should render correct components when patient is in no medical teams', async () => {
@@ -121,7 +108,6 @@ describe('Patient dashboard for HCP', () => {
     expect(dashboard.getByText('Patient statistics')).toBeVisible()
     expect(dashboard.getByText('Device Usage')).toBeVisible()
 
-    expect(dashboard.queryByTestId('remote-monitoring-card')).not.toBeInTheDocument()
     expect(dashboard.queryByTestId('medical-files-card')).not.toBeInTheDocument()
     expect(dashboard.queryByTestId('monitoring-alert-card')).not.toBeInTheDocument()
     expect(dashboard.queryByTestId('chat-card')).not.toBeInTheDocument()
